NiFi: organized non-processor scripts.
