+++ nifi/user_python_extensions/convert_avro_binary_field_to_base64.py
@@ -20,7 +20,7 @@ from nifiapi.properties import (
 from nifiapi.relationship import Relationship
 from overrides import overrides
 from py4j.java_gateway import JavaObject, JVMView
-from utils.helpers.base_nifi_processor import BaseNiFiProcessor
+from utils.nifi.base_nifi_processor import BaseNiFiProcessor
 
 
 class CogStackConvertAvroBinaryRecordFieldToBase64(BaseNiFiProcessor):
