+++ nifi/user_python_extensions/convert_json_record_schema.py
@@ -12,7 +12,7 @@ from nifiapi.properties import ProcessContext, PropertyDescriptor, StandardValid
 from nifiapi.relationship import Relationship
 from overrides import overrides
 from py4j.java_gateway import JavaObject, JVMView
-from utils.helpers.base_nifi_processor import BaseNiFiProcessor
+from utils.nifi.base_nifi_processor import BaseNiFiProcessor
 
 
 class ConvertJsonRecordSchema(BaseNiFiProcessor):
