+++ nifi/user_python_extensions/parse_service_response.py
@@ -14,7 +14,7 @@ from nifiapi.properties import (
 from nifiapi.relationship import Relationship
 from overrides import overrides
 from py4j.java_gateway import JavaObject, JVMView
-from utils.helpers.base_nifi_processor import BaseNiFiProcessor
+from utils.nifi.base_nifi_processor import BaseNiFiProcessor
 
 
 class CogStackParseCogStackServiceResult(BaseNiFiProcessor):
