+++ nifi/user_scripts/cogstack_cohort_generate_random_data.py
@@ -5,7 +5,7 @@ import sys
 from collections import Counter, defaultdict
 from datetime import datetime
 
-from utils.ethnicity_map import ethnicity_map
+from utils.data.ethnicity_map import ethnicity_map
 
 ANNOTATION_DOCUMENT_ID_FIELD_NAME = "meta.docid"
