+++ nifi/user_scripts/legacy_scripts/annotation_creator.py
@@ -3,7 +3,7 @@ import os
 import sys
 import traceback
 
-from utils.sqlite_query import check_db_exists, connect_and_query, create_db_from_file
+from utils.db.sqlite_query import check_db_exists, connect_and_query, create_db_from_file
 
 ANNOTATION_DB_SQL_FILE_PATH = "/opt/cogstack-db/sqlite/schemas/annotations_nlp_create_schema.sql"
