+++ nifi/user_scripts/utils/codecs/cerner_blob.py
