+++ nifi/user_scripts/utils/serialization/avro_json_encoder.py
